+++ rakaia/callbacks/pixel.py
@@ -69,7 +69,7 @@ from rakaia.utils.graph import strip_invalid_shapes_from_graph_layout
 from rakaia.inputs.loaders import (
     previous_roi_trigger,
     next_roi_trigger,
-    adjust_option_height_from_list_length, set_roi_tooltip_based_on_length)
+    adjust_option_height_from_list_length, set_roi_tooltip_based_on_length, valid_key_trigger, mask_toggle_trigger)
 from rakaia.callbacks.pixel_wrappers import parse_global_filter_values_from_json, parse_local_path_imports, \
     mask_options_from_json, bounds_text, generate_annotation_list, no_json_db_updates
 from rakaia.io.session import (
@@ -1878,6 +1878,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         raise PreventUpdate
 
     @dash_app.callback(Output('data-collection', 'value', allow_duplicate=True),
+                       Output('apply-mask', 'value', allow_duplicate=True),
                        Input('prev-roi', 'n_clicks'),
                        Input('next-roi', 'n_clicks'),
                        Input('keyboard-listener', 'event'),
@@ -1888,24 +1889,28 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('region-annotation-modal', 'is_open'),
                        State('main-tabs', 'active_tab'),
                        State('tour_component', 'isOpen'),
+                       State('apply-mask', 'value'),
                        prevent_initial_call=True)
-    def click_to_new_roi(prev_roi, next_roi, key_listener, n_events, cur_data_selection, cur_options,
-                         allow_arrow_change, annotating_region, active_tab, open_tour):
+    def use_key_listener(prev_roi, next_roi, key_listener, n_events, cur_data_selection, cur_options,
+                         allow_arrow_change, annotating_region, active_tab, open_tour, mask_stat):
         """
-        Use the forward and backwards buttons to click to a new ROI
-        Alternatively, use the directional arrow buttons from an event listener
+        Use the key event listener to trigger the following actions:
+            - Use the forward and backwards buttons to click to a new ROI
+            - Alternatively, use the directional arrow buttons from an event listener
+            - Use the arrow up button to toggle on/off the mask
         """
-        if None not in (cur_data_selection, cur_options) and not (ctx.triggered_id == 'keyboard-listener' and
-        not allow_arrow_change) and not annotating_region and active_tab == 'pixel-analysis' and not open_tour:
+        if None not in (cur_data_selection, cur_options) and not (ctx.triggered_id == 'keyboard-listener' and not allow_arrow_change) and \
+            not annotating_region and active_tab == 'pixel-analysis' and not open_tour and valid_key_trigger(key_listener):
             cur_index = cur_options.index(cur_data_selection)
+            mask_change = not mask_stat if mask_toggle_trigger(ctx.triggered_id, key_listener, n_events) else dash.no_update
             try:
                 prev_trigger = previous_roi_trigger(ctx.triggered_id, prev_roi, key_listener, n_events)
                 next_trigger = next_roi_trigger(ctx.triggered_id, next_roi, key_listener, n_events)
                 if prev_trigger and cur_index != 0:
-                    return cur_options[cur_index - 1] if cur_options[cur_index - 1] != cur_data_selection else dash.no_update
+                    return cur_options[cur_index - 1] if cur_options[cur_index - 1] != cur_data_selection else dash.no_update, mask_change
                 elif next_trigger:
-                    return cur_options[cur_index + 1] if cur_options[cur_index + 1] != cur_data_selection else dash.no_update
-                else: raise PreventUpdate
+                    return cur_options[cur_index + 1] if cur_options[cur_index + 1] != cur_data_selection else dash.no_update, mask_change
+                else: return dash.no_update, mask_change
             except IndexError: raise PreventUpdate
         raise PreventUpdate