+++ rakaia/plugins/__init__.py
@@ -31,30 +31,34 @@ class PluginModes:
     # Each should return a modified quantification sheet
     @staticmethod
     def leiden(quantification_results: Union[dict, go.Figure], input_category: str,
-                  output_category: str):
+                  output_category: str, **kwargs):
         """
         Generate leiden clustering
 
         :param quantification_results: `pd.DataFrame` or `anndata.AnnData` object containing tabular object intensity measurements
         :param input_category: Placeholder argument for similar plugins. NOt used for leiden clustering
         :param output_category: Column to store the outputs of the clustering procedure
+        :param kwargs: Keyword arguments to pass to `scanpy.tl.leiden`
+
         :return: Quantification results with the cluster labels added under the output column.
         """
-        return leiden_clustering(quantification_results, output_category)
+        return leiden_clustering(quantification_results, output_category, **kwargs)
 
     @staticmethod
     def random_forest(quantification_results: Union[dict, go.Figure], input_category: str,
-                  output_category: str):
+                  output_category: str, **kwargs):
         """
         Train and predict using a rando forest classifier
 
         :param quantification_results: `pd.DataFrame` or `anndata.AnnData` object containing tabular object intensity measurements
         :param input_category: Existing annotation column in the object that specifies the classes used to fit the model
         :param output_category: Column to store the output labels of the classifier prediction
+        :param kwargs: Keyword arguments to pass to `RandomForestClassifier`
+
         :return: Quantification results with the prediction labels added under the output column.
         """
         return QuantificationRandomForest(
-            quantification_results, input_category, output_category).quantification_with_labels(True)
+            quantification_results, input_category, output_category, **kwargs).quantification_with_labels(True)
 
 
 class PluginModelModes:
@@ -68,7 +72,7 @@ class PluginModelModes:
     random_forest = partial(PluginModes.random_forest)
 
 def run_quantification_model(quantification_results: Union[dict, go.Figure], input_category: Union[str, None]=None,
-                             output_category: str="out", mode: str="leiden"):
+                             output_category: str="out", mode: str="leiden", **kwargs):
     """
     Run a quantification model on a set of objects with quantified channel intensities. Requires a specific model
     mode provided by the user
@@ -78,5 +82,5 @@ def run_quantification_model(quantification_results: Union[dict, go.Figure], inp
         raise PluginNotFoundError(f"The plugin mode provided: {mode} is not a supported plugin."
                                   f"Current plugins include: {PluginDescriptors.descriptors}")
     mode = mode.replace(" ", "_") if mode else ""
-    with_model = getattr(PluginModelModes, mode)(quantification_results, input_category, output_category)
+    with_model = getattr(PluginModelModes, mode)(quantification_results, input_category, output_category, **kwargs)
     return with_model